Extract shareable methods into separate module
