+++ lib/api/v3/work_packages/work_packages_api.rb
@@ -27,9 +27,7 @@
 #++
 
 require 'api/v3/activities/activity_representer'
-require 'api/v3/work_packages/work_package_contract'
 require 'api/v3/work_packages/work_package_representer'
-require 'api/v3/work_packages/form/work_package_payload_representer'
 
 module API
   module V3
@@ -40,59 +38,13 @@ module API
             requires :id, desc: 'Work package id'
           end
           route_param :id do
+            helpers WorkPackagesSharedHelpers
             helpers do
               attr_reader :work_package
 
               def work_package_representer
                 WorkPackageRepresenter.create(@work_package,
                                               current_user: current_user)
-              end
-
-              def write_work_package_attributes
-                if request_body
-                  begin
-                    # we need to merge the JSON two times:
-                    # In Pass 1 the representer only has custom fields for the current WP type
-                    # After Pass 1 the correct type information is merged into the WP
-                    # In Pass 2 the representer is created with the new type info and will be able
-                    # to also parse custom fields successfully
-                    merge_json_into_work_package!(request_body.to_json)
-                    merge_json_into_work_package!(request_body.to_json)
-                  rescue ::API::Errors::Form::InvalidResourceLink => e
-                    fail ::API::Errors::Validation.new(e.message)
-                  end
-                end
-              end
-
-              # merges the given JSON representation into @work_package
-              def merge_json_into_work_package!(json)
-                payload = Form::WorkPackagePayloadRepresenter.create(
-                  @work_package,
-                  enforce_lock_version_validation: true)
-                payload.from_json(json)
-              end
-
-              def request_body
-                env['api.request.body']
-              end
-
-              def write_request_valid?
-                contract = WorkPackageContract.new(@work_package, current_user)
-
-                contract_valid = contract.validate
-                represented_valid = @work_package.valid?
-
-                return true if contract_valid && represented_valid
-
-                # We need to merge the contract errors with the model errors in
-                # order to have them available at one place.
-                contract.errors.keys.each do |key|
-                  contract.errors[key].each do |message|
-                    @work_package.errors.add(key, message)
-                  end
-                end
-
-                false
               end
             end
